FruitCounter_Project_GH-Actions/requirements.txt
kivy
numpy
Pillow
tensorflow
